Temporal sync calibration + subpixel PCCR + push-based capture
- True temporal sync: for each eye frame at time T, fetch the world cam
  frame at T via new GET /closest_frame endpoint, run ArUco on that frame,
  compute scene_xy from its homography — PCCR, ArUco, and target position
  all share the same timestamp
- Add GET /closest_frame?cam=<id>&ts_ms=<float> to receiver.py using the
  existing _recent ring buffer (500ms, 200ms sync window)
- Subpixel pupil/glint centroids: change PupilResult/GlintResult centers
  from int to float to eliminate frozen PCCR (was causing ~49% of fixations
  to be skipped as std=0)
- Push-based calibration: receiver fires _eye_analysis_thread per frame,
  pushes PCCR with camera capture timestamp to calibration server; replaces
  polling-based architecture
- Engine returns PCCR in response headers (X-Pccr-Dx/Dy/Ts) to avoid a
  second /result round-trip
- CLAHE preprocessing in ArUco detection for low-light frames
- IQR outlier rejection and frozen-sample detection per fixation

+++ compute/glint_detector.py
@@ -19,8 +19,8 @@ from dataclasses import dataclass, field
 
 @dataclass
 class GlintResult:
-    glints: list[tuple[int, int]] = field(default_factory=list)   # all detected glints
-    primary: tuple[int, int] | None = None                        # best glint for PCCR
+    glints: list[tuple[float, float]] = field(default_factory=list)  # all detected glints
+    primary: tuple[float, float] | None = None                       # best glint for PCCR
     debug_mask: np.ndarray | None = None
     intermediate_frames: dict = field(default_factory=dict)
 
@@ -96,8 +96,8 @@ class GlintDetector:
             M = cv2.moments(cnt)
             if M["m00"] == 0:
                 continue
-            gx = int(M["m10"] / M["m00"])
-            gy = int(M["m01"] / M["m00"])
+            gx = M["m10"] / M["m00"]
+            gy = M["m01"] / M["m00"]
 
             # Peak brightness in the blob region
             blob_mask = np.zeros_like(gray)
